enable window communication
the app ist listening on a `connectionSuccessFul` event on the window. When the event arrives it will set the session with the creds that are attached to the event. the event will be fired e.g. by the native-electron app.
It is also listening on messages of the opened modal when you click on connect. The popup can post a message to its parent window. When the message arrives it will set the session